Fix for expand-less copies with multiple files
Introduces a new test that uses a cwd- and expand-less target with
multiple input files being copied to a single destination directory. In
the task itself, the loop that operates on the list of `src` paths no
longer persists the `dest` from loop-to-loop, allowing the `src`
filename to actually show up in the full `dest` for successive files.

Fixes #220.
Partially reverts a15feecc7763269e995235dcb4647b223d67f516.